Move Pv watcher code to individual library
